# src/api/openapi.yaml
...
openapi: 3.0.0
info:
  title: Insights Remediations
  description: Insights Remediations Service
  version: 1.0.0
  contact:
    email: jozef@redhat.com
servers:
  - url: /r/insights/platform/remediations/v1
...
paths:
...
  /resolutions/{issue}:
    get:
      tags:
        - resolutions
      summary: Resolution metadata
      description: Provides information about resolutions available for the given issue
      operationId: resolutions
      parameters:
        - $ref: '#/components/parameters/IssueId'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Resolutions'
        400:
          $ref: '#/components/responses/BadRequest'
        404:
          description: Issue not known or Ansible resolution not available
  /resolutions:
    post:
      tags:
        - resolutions
      summary: Resolution metadata (batch)
      description: Provides information about resolutions available for the given issues
      operationId: resolutionsBatch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolutionsBatchInput'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolutionsBatch'
        400:
          $ref: '#/components/responses/BadRequest'
...
components:
  parameters:
    IssueId:
      in: path
      name: issue
      description: Issue identifier (e.g. `vulnerabilities:CVE_2017_6074_kernel|KERNEL_CVE_2017_6074`)
      required: true
      schema:
        $ref: '#/components/schemas/IssueId'
...
  schemas:
...
    IssueId:
      type: string
      pattern: ^(advisor|vulnerabilities|compliance|test):[\w\d-_|:\.]+$
      example: vulnerabilities:CVE_2017_6074_kernel|KERNEL_CVE_2017_6074
...
    ResolutionNeedsReboot:
      type: boolean
      description: Indicates whether the given resolution involves system reboot
    ResolutionRisk:
      type: integer
      enum:
        - -1
        - 1
        - 2
        - 3
        - 4
      example: 2
    ResolutionType:
      type: string
      example: fix
...
    ResolutionsBatchInput:
      type: object
      additionalProperties: false
      required: [issues]
      properties:
        issues:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/IssueId'
      example:
        issues:
          - vulnerabilities:CVE-2018-1087
          - vulnerabilities:CVE_2017_6074_kernel|KERNEL_CVE_2017_6074377300
          - advisor:non-existent-issue
...
    Error:
      type: object
      additionalProperties: false
      required: [errors]
      properties:
        errors:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [id, status, code, title]
            properties:
              id:
                type: string
              status:
                type: integer
                example: 400
              code:
                type: string
                example: UNKNOWN_SYSTEM
              title:
                type: string
                example: Unknown system identifier 'abcd'

    # Resolution specific types
    Resolutions:
      type: object
      additionalProperties: false
      required: [id, resolution_risk, resolutions]
      properties:
        id:
          $ref: '#/components/schemas/IssueId'
        resolution_risk:
          $ref: '#/components/schemas/ResolutionRisk'
        resolutions:
          type: array
          items:
            type: object
            additionalProperties: false
            required: [id, description, needs_reboot, resolution_risk]
            properties:
              id:
                $ref: '#/components/schemas/ResolutionType'
              description:
                type: string
                example: Disable DCCP kernel module
              needs_reboot:
                $ref: '#/components/schemas/ResolutionNeedsReboot'
              resolution_risk:
                $ref: '#/components/schemas/ResolutionRisk'
    ResolutionsBatch:
      type: object
      example:
        vulnerabilities:CVE-2018-1087:
          id: vulnerabilities:CVE-2018-1087
          resolution_risk: -1
          resolutions:
            - description: Upgrade packages affected by CVE-2018-1087
              id: fix
              needs_reboot: true
              resolution_risk: -1
        vulnerabilities:CVE_2017_6074_kernel|KERNEL_CVE_2017_6074:
          id: vulnerabilities:CVE_2017_6074_kernel|KERNEL_CVE_2017_6074
          resolution_risk: 3
          resolutions:
            - description: Update system to the latest kernel and reboot
              id: kernel_update
              needs_reboot: true
              resolution_risk: 3
            - description: Disable DCCP kernel module
              id: mitigate
              needs_reboot: true
              resolution_risk: 3
            - description: Make sure SELinux is enabled, enforcing and has selinux-policy-3.13.1-81.el7 or later on RHEL7
              id: selinux_mitigate
              needs_reboot: true
              resolution_risk: 3
        advisor:non-existent-issue: false


    # Remediation specific types
...
  responses:
...
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...